Stackhpc upstream master (#5)
* Add config option to not collect logs

Collecting logs from BMCs can be very slow, so you may wish to
disable doing so. However, the logs can still be useful.

This change adds support for optionally overriding the `collectlogs`
config setting. One use case envisaged would be to set collectlogs
to false in the config, allowing the BMC to be scraped for data
every few minutes. Then every hour or so, the logs could be collected
via the `collectlogs` override.

---------

Co-authored-by: Matt Anson <matta@stackhpc.com>