Granular control for metric collection
We see an issue where a BMC is reporting 42+ ethernet
interfaces and not returning any data for them. This is likely
a bug in the BMC FW.

This patch allows skipping the collection of certain groups
of metrics to work around the bug. It adds an optional
configuration section which can be used to control what is
collected.